Added core functions to the wallpapers to check if items are deleted locally.
This commit adds in a very specific function, check the wallpapers you have downloaded
and make sure they are there befor showing an apply icon. So if you go into the gallery
and delete a wallpaper, the app will check and see its gone and remove it from the
local db where the cached entry is.